+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.2.0] - 2021-11-03
+### Added
+- Added a new script to go into the running game container.
+
+### Changed
+- The game start and stop were split up two separated scripts.
+
+
 ## [0.1.0] - 2021-10-26
 ### Added
 - Added makefile to manage the project.